wip try rerunning mkstringprep
After the commit that first added this, 2548c730c in 2003, the only
changes to it appear to have been to adjust to changes in Python.

Here's what I get on attempting a rerun of the script.  There's a
bunch of whitespace change -- and also some change in the actual data!
Not sure what changed to cause that since the last time this script
was rerun (whenever indeed that was.)

It's not 100% clear to me how to reproduce; here's what I did:

 $ curl -s https://tools.ietf.org/rfc/rfc3454.txt >rfc3454.txt
 $ ./python Tools/unicode/mkstringprep.py >Lib/stringprep.py

Here

 $ sha256sum rfc3454.txt
 eb722fa698fb7e8823b835d9fd263e4cdb8f1c7b0d234edf7f0e3bd2ccbb2c79  rfc3454.txt