tupleobject: Make the freelist unconditional.
This optimization was added in 1993 (commit 842d2ccdc), and we've
been running with it ever since.  I think at this point it's fair
to say that it's no longer experimental.  That lets us simplify
the code by removing a number of `#ifdef`s.

Also remove the `#ifndef` guards around setting the parameters of
the freelist.  The only way those could be caused to do anything
(which is also the only way of disabling the optimization) seems
to be to manually add definitions of `PyTuple_MAXSAVESIZE` or
`PyTuple_MAXFREELIST` elsewhere, as there doesn't appear to be
any kind of configuration knob for them; and neither one was ever
documented.