WIP fix linebreaks to match Unicode
See https://bugs.python.org/issue7643 .

Seems pretty clear that the actual Unicode prescription here
is that 001C..001E don't belong in the set of line breaks.

In _PyUnicode_Init, note that the list was wrong in its own way:
it left out 000B..000C.  This didn't cause a bug because it turns
out we never consult the filter for values that fit in ASCII.
So, note that in a comment and also cut all the other entries that
don't need to be there.