wip absurdly rice escape-decoding
For the `switch` statement, GCC produces a jump table.  It's a little
over 400 bytes long, and then the code it jumps to is spread over a
few hundred more.

How about making our own table that fits in a cache line, and having
just a few lines of code to act on it?