Fix asserts to actually work in two spots where Py_TOUPPER had caused confusion.
These didn't actually work -- if e.g. this code were to through some
bug end up processing data with '\u066c' ARABIC THOUSANDS SEPARATOR
in it, it'd just mask it to '\x6c' aka 'l', uppercase to 'L', and
cheerfully stick that in the output.

With the assert first, on the other hand, if I artificially
s/toupper/1/ to make the latter block run when it shouldn't, then
`test_int__format__locale` in test_types.py hits the assert.