no CHARMASK when calling function
In this context Py_CHARMASK has no effect on behavior -- we're about
to implicitly convert to char anyway, which will mask down to the low
bits just like Py_CHARMASK would.

Its only effect is to potentially suppress a warning (under
-Wconversion, which is not one of our standard warning flags), by
looking like an explicit cast.  And if the compiler does see cause for
such a warning, we really only want to suppress it with an explicit
cast right there in the calling code, where it's possible for a reader
to evaluate whether there's some reason to be confident it's fine.