WIP DRAFT Indent all ifdefs everywhere.
Almost surely too much to merge in one go, even once tweaked as
needed, even provided there's agreement on doing this at all.

This indentation really helps a ton in understanding the structure of
the code when the nesting is complex, or when it extends across long
stretches (other than when it's the whole file.)

Done automatically like so:

  $ perl -i -lpe '
        $indentstr = "  " x $indent;
        s/^#\s*/#${indentstr}/;
        s/^#\s*\K  (el|endif)/$1/;
        $indent += 1 if (/^#\s*if/);
        $indent -= 1 if (/^#\s*endif/);
    ' $(git ls-files **/*.[ch] | grep -v /clinic/)

TODO:
* Exclude generated files.  The `grep -v /clinic/` is a start;
  better, rely on gitattributes.
* Exclude include-guards that (are in a style that's meant to)
  cover an entire file.

Why two spaces?  For one, it's closest to the status quo.  Here's
master:

  $ git grep -hE '^#\s*' **/*.[ch] \
    | perl -lpe 's/^#\s*\K.*/|/' \
    | sort | uniq -c
        4 #			|
       22 #		|
      106 #	|
        8 #                               |
       15 #                       |
        9 #               |
        2 #           |
        1 #         |
       20 #        |
       61 #       |
        1 #      |
       52 #    |
       86 #   |
      479 #  |
      292 # |
    23730 #|

So there are 479 two-space-indented lines vs. 52 four-space-indented.
The latter are actually outnumbered by 106 tab-indented.

Here's after this change:

  $ git grep -hE '^#\s*' **/*.[ch] \
    | perl -lpe 's/^#\s*\K.*/|/' \
    | sort | uniq -c
        1 #                  |
        2 #                |
        2 #              |
        2 #            |
       15 #          |
       64 #        |
      279 #      |
     1600 #    |
     5349 #  |
    17557 #|