utime: begin reducing explosion of codepaths slightly
Specifying either follow_symlinks or dir_fd must be an error if
MS_WINDOWS -- I believe it is already, but if it isn't, that's a bug
because they certainly won't work: we don't look at them in that case.

That in turn means all the more complex conditions won't be reached.

TODO next steps: try to reduce complexity of N combinations of
arguments x M combinations of HAVE_*.