ifdef -> assert: utime_dir_fd
Just like the parent, if it were possible to reach this assert then
we'd already have a bug.

And in fact if defined(HAVE_FUTIMESAT) || defined(HAVE_UTIMENSAT) is
false, then we use dir_fd_unavailable to reject a non-default dir_fd.
Which means that if this condition could only be true if
!follow_symlinks... but in that case the previous condition would have
been true.