utime: Complete making checks parallel logic.
Not much to say in the remaining cases; just briefly comment why.

Move the checks for platform-independent preconditions we've chosen
for our API above the checks for platform-specific availability of
features.  This means that if e.g. you pass `dir_fd` and also a path
for `fd`, you'll always get an error message about that even if you're
also on a platform where one or both of those features isn't available
anyway.  That seems most appropriate.