clang-tidy: add crude sketch of a way to run it
Like the logic in the parent, this needs some assimilating into the
config system.

Having seen the results, I'm also not sure this turns out to be useful.
Probably the best suggestions are those of `bugprone-macro-parentheses`;
but it turns out we have a *ton* of macro definitions that don't use
parens, and I'm not sure that's a particularly frequent cause of bugs.

There are a bunch more checks I haven't tried yet.  Full list here:
  https://clang.llvm.org/extra/clang-tidy/checks/list.html
or with `clang-tidy -checks=* -list-checks`.  So far I've tried:

 * those enabled by default, aka `clang-analyzer-*` -- and only tried
   that at small scale and not on the whole codebase

 * those in `bugprone-*`