fix another, in _ssl
../Modules/_ssl.c:959:20: warning: Access to field 'bio' results in a dereference of a null pointer (loaded from variable 'inbio')
        BIO_up_ref(inbio->bio);
                   ^~~~~~~~~~

This path can't currently happen, because it's effectively part of the
interface to this function that either `sock` is non-NULL, or both
`inbio` and `outbio` are.  The function is static, and its two
call-sites adhere to that interface, so there's no live bug here.

But: that precondition in the interface isn't written down in any way.
Each of these parameters can individually be NULL, so the mutual
exclusion is a bit of a subtle fact.

For a simple fix, just make sure the function returns promptly with an
error if we somehow end up passing it an invalid combination of this
kind.