cut a stray assignment in _ssl
../Modules/_ssl.c:4179:13: warning: Value stored to 'ok' is never read
            ok = 0;
            ^    ~

Indeed this is redundant with the `goto error` just below.

In all the other error paths in this function, we rely on `goto error`
to take care of setting `ok` to 0.  Given that there's no substantive
difference accomplished by doing it differently in this case, it's
clearest if it's done the same way so the reader doesn't wonder if
there's a subtle difference they're just not seeing.