Cut a dead store.
../Python/pylifecycle.c:683:5: warning: Value stored to 'config' is never read
    config = &tstate->interp->config;
    ^        ~~~~~~~~~~~~~~~~~~~~~~~

I haven't stared at this area of the code enough to tell if this is a
bug where something else was intended, or if the right fix is just to
delete this line.

Looks like this has been this way since commit 6d43f6f08 earlier this
year.  There's been some further refactoring around here since then,
but the logic in this function appears substantially unchanged.