fix what appears to be always a null-deref
NOTE: Next tip for higher-quality results: build in debug mode,
  with assertions.

---

Got the story from the Clang static analyzer, like so:

  $ make Modules/_struct.o CFLAGS='--analyze --analyzer-output html'

(The CLI is kind of a mess -- that ends up making Modules/_struct.o
a *directory*, with an HTML file in it.  But the information in the
HTML is great stuff.)

Looking, I'm pretty sure that with the existing callers, in fact
`*ptr` is *always* NULL if we get to this line.

This function doesn't decref `*ptr` in its other paths; so it
shouldn't here either.  Just delete the line.