Fix a condition that would overflow if it could fail.
This condition is always true: LLONG_MAX is always at least 2**63 - 1,
as a requirement of C99.

Moreover, in the constant expression controlling a preprocessor
conditional, all signed integers "act as if they have the same
representation as", as the standard puts it, the integer type
`intmax_t`; and it's uncommon to have any integer types larger than
`long long`.  That means that even in a hypothetical variation of
this code where the left-hand side of the inequality were actually
larger than LLONG_MAX... that would typically just mean that it
overflowed.