bpo-38043: Fix wrong results for unicodedata.ucd_3_2_0.is_normalized.
This bit of logic was fairly suspicious-looking if you looked closely
at it together with its comment: if quickchecks must be disabled, why
are we returning a definite NO?  Disabling quickcheck seems like it
should mean we return MAYBE, i.e. "quickcheck can't tell".

And, in fact, adding a few smoke-tests shortly finds `is_normalized`
on the old database was always returning False, even for strings
like "A".

The root cause is basically the same as for bpo-37966.  This helper
`is_normalized_quickcheck` started life with the `normalized`
function as its only caller; and although it was described as
implementing the quickcheck algorithm from the standard, in fact it
differed in some ways that simply didn't matter for that caller,
which treats NO and MAYBE the same.

Fix the bug, and add those smoke-tests.  Not a bug many people are
likely to actually hit -- ucd_3_2_0 likely has very few users
outside IDNA implementations like the one in the stdlib -- but good
to fix if only for the sake of people reading this code.