handle all aliases, not just first
Among these three properties, as of Unicode 12.1.0 this applies only
to a few of the General_Category values.

It wouldn't be terrible to do without the additional aliases for
those; but it's better to follow the upstream data model.  There may
be more aliases in the future, plus some other properties like Block
have aliases like "Latin_1" (for "Latin_1_Sup") as non-first, which
would be annoying to miss.