And add to ucd_3_2_0.
This doesn't attempt to actually record the older version of this
data.  Rather, this is here just as a compatibility convenience.
Given that the applications of the old data are quite specific and
haven't needed this information yet, it seems unlikely anyone will
ever really want to know from this module just which aliases existed
already in 3.2.0 and which didn't yet.

(Alternatively, it might be appropriate to just leave these attributes
off of ucd_3_2_0 entirely.)